feat: 중간 테이블인 이슈레이블, 이슈담당자 엔티티와 복합키 생성
