+++ backend/src/main/java/codesquad/team01/issuetracker/issue/domain/IssueLabelId.java
@@ -1,2 +1,13 @@
-package codesquad.team01.issuetracker.issue.domain;public class IssueLabelId {
+package codesquad.team01.issuetracker.issue.domain;
+
+import java.io.Serializable;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+
+@EqualsAndHashCode
+@AllArgsConstructor
+public class IssueLabelId implements Serializable {
+	private Integer issueId;
+	private Integer labelId;
 }
